requirements.txt
asgiref==3.6.0
attrs==22.2.0
chainer==7.8.1
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
drf-spectacular==0.25.1
drf-spectacular-sidecar==2023.3.1
filelock==3.9.0
inflection==0.5.1
install==1.3.5
jsonschema==4.17.3
numpy==1.24.2
protobuf==4.22.0
psycopg2==2.9.5
PyJWT==2.6.0
pymongo==4.3.3
pypandoc==1.10
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
room==0.1.11
six==1.16.0
sqlparse==0.4.3
tornado==6.2
typing_extensions==4.5.0
uritemplate==4.1.1
